Shanmukh/requirements.txt
streamlit==1.39.0
tensorflow-cpu==2.16.1
pandas==2.2.2
numpy==1.26.4
pillow==10.3.0
